Update ezTime.cpp
The Function makeOrdinalTime does not work properly if you set the ordinal to 5 (Last dow in month and year) because you check if the value w is equal to 0 for the "Last week" rule. This will never happen. With the check if (w == 5) or  if (ordinal ==0) the code works perfectly.
This change fixes the issue #38