fix bug where timezones not observing DST were incorrectly translated
Timezones like Pacific/Honolulu and America/Phoenix (and probably others) do not observe DST.  They always stay on their standard time.

The previous code essentially treated all calls to `tzTime` in this time zone like the timestamp `t` was a UTC timestamp (subtracting the the offset).  However, this is incorrect and should be able to translate both ways depending on what `local_or_utc` is set to.